backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pillow>=10.2.0
redis>=5.0.1
celery>=5.3.6
aiofiles>=23.2.1
httpx>=0.26.0
python-dotenv>=1.0.0
slowapi>=0.1.9
loguru>=0.7.2
cloudinary>=1.38.0
websockets>=12.0
greenlet>=3.0.3
email-validator>=2.1.0
